backend/requirements.txt
django==5.0.6
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.4.0
python-dotenv==1.0.1
psycopg[binary]==3.1.19
psycopg_pool==3.2.4
msgpack==1.0.8
channels==4.1.0
channels-postgres==1.1.2
uvicorn==0.35.0
daphne==4.1.2
django==5.0.6

